app/requirements.txt
aiohttp
python-dotenv
discord.py
fastapi
uvicorn
ollama